+++ ncdump/tst_formatx3.sh
@@ -21,7 +21,7 @@ echo "*** Fail: extended format for a classic file"
 ECODE=1
 fi
 
-echo "Test extended format output for a 64-bit netcdf-3 file"
+echo "Test extended format output for a 64-bit offset netcdf-3 file"
 rm -f tmp
 ../ncgen/ncgen -k nc6 -b -o ./test.nc $srcdir/ref_tst_small.cdl
 ./ncdump -K test.nc >tmp
@@ -34,7 +34,7 @@ echo "Test extended format output for a 64-bit CDF-5 classic file"
 rm -f tmp
 ../ncgen/ncgen -k5 -b -o ./test.nc $srcdir/ref_tst_small.cdl
 ./ncdump -K test.nc >tmp
-if ! fgrep '64-bit data mode=00000010' <tmp ; then
+if ! fgrep '64-bit data mode=00008000' <tmp ; then
 echo "*** Fail: extended format for a 64-bit CDF-5 classic file"
 ECODE=1
 fi